refactor: standardize id params to int (A5)
Aligned stray 'long' id parameters with the int entity ids (Child/Purchase/Table
services + controllers, ChildUpdateDto, UnifiedSearchResultDto); removed (int) casts.
Harness 10/10. (DB id cols are bigint; entity ids are int — fine at current scale.)

+++ Controllers/ChildController.cs
@@ -26,7 +26,7 @@ namespace Iroh.Controllers
         }
 
         [HttpPost("customers/{parentId}")]
-        public async Task<IActionResult> Create(long parentId, [FromBody] ChildCreateDto dto)
+        public async Task<IActionResult> Create(int parentId, [FromBody] ChildCreateDto dto)
         {
             // Sistem misafiri → servis BusinessRuleException atar → handler 400.
             var child = await _childService.CreateChild(parentId, dto.name, dto.birthDate);
@@ -34,7 +34,7 @@ namespace Iroh.Controllers
         }
 
         [HttpGet("customers/{parentId}")]
-        public async Task<IActionResult> GetByParent(long parentId)
+        public async Task<IActionResult> GetByParent(int parentId)
         {
             var children = await _childService.GetChildrenByParentId(parentId);
             return Ok(ApiResponse.Ok(children, "Başarılı"));
@@ -48,7 +48,7 @@ namespace Iroh.Controllers
         }
 
         [HttpDelete("{id}")]
-        public async Task<IActionResult> Delete(long id)
+        public async Task<IActionResult> Delete(int id)
         {
             await _childService.DeleteChild(id);
             return Ok(ApiResponse.Ok<object?>(null, "Çocuk başarıyla silindi"));
